requirements.txt
Django==6.0.1
gunicorn==23.0.0
google-generativeai==0.8.6
pdf2image==1.17.0
python-dotenv==1.2.1
pillow==12.1.0
requests==2.32.5